+++ asterisk_click2dial/__init__.py
@@ -1,4 +1,4 @@
 # -*- coding: utf-8 -*-
 
-from . import asterisk_click2dial
+from . import models
 from . import controller
