+++ asterisk_click2dial/__manifest__.py
@@ -4,7 +4,7 @@
 
 {
     'name': 'Asterisk Click2dial',
-    'version': '9.0.1.0.0',
+    'version': '10.0.1.0.0',
     'category': 'Phone',
     'license': 'AGPL-3',
     'summary': 'Asterisk-Odoo connector',
@@ -80,14 +80,14 @@ http://www.akretion.com/products-and-services/openerp-asterisk-voip-connector
     'depends': ['base_phone'],
     'external_dependencies': {'python': ['Asterisk']},
     'data': [
-        'asterisk_server_view.xml',
-        'res_users_view.xml',
+        'views/asterisk_server.xml',
+        'views/res_users.xml',
         'security/ir.model.access.csv',
         'web_asterisk_click2dial.xml',
         ],
-    'demo': ['asterisk_click2dial_demo.xml'],
+    'demo': ['demo/asterisk_click2dial_demo.xml'],
     'qweb': ['static/src/xml/*.xml'],
     'css': ['static/src/css/*.css'],
     'application': True,
-    'installable': False,
+    'installable': True,
 }
