+++ asterisk_click2dial/models/__init__.py
@@ -1,0 +1,5 @@
+# -*- coding: utf-8 -*-
+
+from . import asterisk_server
+from . import res_users
+from . import phone_common
