+++ asterisk_click2dial/static/src/xml/asterisk_click2dial.xml
@@ -15,10 +15,12 @@
     </t>
 
     <!-- Add Dial button in phone widget for click2dial feature -->
+    <!--
     <t t-extend="FieldPhone">
         <t t-jquery="a.oe_form_uri" t-operation="after">
             <a id="click2dial" href="#" class="oe_bold"/>
         </t>
     </t>
+    -->
 
 </template>
