--- crm_claim_phone/crm_claim_view.xml
@@ -1,22 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<!--
-  © 2014-2016 Akretion (Alexis de Lattre <alexis.delattre@akretion.com>)
-  License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
--->
-
-<odoo>
-<data>
-
-<record id="crm_case_claims_form_view" model="ir.ui.view">
-    <field name="name">crm_claim_phone.crm_claim.form</field>
-    <field name="model">crm.claim</field>
-    <field name="inherit_id" ref="crm_claim.crm_case_claims_form_view"/>
-    <field name="arch" type="xml">
-        <field name="partner_phone" position="attributes">
-            <attribute name="widget">phone</attribute>
-        </field>
-    </field>
-</record>
-
-</data>
-</odoo>
