+++ crm_phone/models/__init__.py
@@ -1,0 +1,7 @@
+# -*- coding: utf-8 -*-
+
+from . import crm_phonecall
+from . import crm_lead
+from . import res_partner
+from . import res_users
+from . import phone_common
