+++ crm_phone/security/phonecall_security.xml
@@ -1,19 +1,18 @@
 <?xml version="1.0" encoding="utf-8"?>
-<odoo>
-<data noupdate="1">
+<odoo noupdate="1">
 
 
 <record id="my_crm_phonecall_rule" model="ir.rule">
     <field name="name">Personal Phone Calls</field>
     <field name="model_id" ref="model_crm_phonecall"/>
-    <field name="groups" eval="[(4, ref('base.group_sale_salesman'))]"/>
+    <field name="groups" eval="[(4, ref('sales_team.group_sale_salesman'))]"/>
     <field name="domain_force">['|', ('user_id', '=', False), ('user_id', '=', user.id)]</field>
 </record>
 
 <record id="all_crm_phonecall_rule" model="ir.rule">
     <field name="name">All Phone Calls</field>
     <field name="model_id" ref="model_crm_phonecall"/>
-    <field name="groups" eval="[(4, ref('base.group_sale_salesman_all_leads'))]"/>
+    <field name="groups" eval="[(4, ref('sales_team.group_sale_salesman_all_leads'))]"/>
     <field name="domain_force">[(1, '=', 1)]</field>
 </record>
 
@@ -24,5 +23,4 @@
 </record>
 
 
-</data>
 </odoo>
