+++ crm_phone/tests/test_crm_phone.py
@@ -1,8 +1,8 @@
 # -*- coding: utf-8 -*-
 # © 2016 Akretion France (Alexis de Lattre <alexis.delattre@akretion.com>)
+# License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
 
-
-from openerp.tests.common import TransactionCase
+from odoo.tests.common import TransactionCase
 
 
 class TestCRMPhone(TransactionCase):
