--- event_phone/event_view.xml
@@ -1,24 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<!--
-  © 2014-2016 Akretion (Alexis de Lattre <alexis.delattre@akretion.com>)
-  License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl)
--->
-
-<odoo>
-<data>
-
-
-<record id="view_event_registration_form" model="ir.ui.view">
-    <field name="name">event_phone.event_registration.form</field>
-    <field name="model">event.registration</field>
-    <field name="inherit_id" ref="event.view_event_registration_form"/>
-    <field name="arch" type="xml">
-        <field name="phone" position="attributes">
-            <attribute name="widget">phone</attribute>
-        </field>
-    </field>
-</record>
-
-
-</data>
-</odoo>
