+++ hr_phone/__manifest__.py
@@ -4,7 +4,7 @@
 
 {
     'name': 'HR Phone',
-    'version': '9.0.1.0.0',
+    'version': '10.0.1.0.0',
     'category': 'Phone',
     'license': 'AGPL-3',
     'summary': 'Validate phone numbers in HR',
@@ -24,10 +24,7 @@ for any help or question about this module.
     'author': "Akretion,Odoo Community Association (OCA)",
     'website': 'http://www.akretion.com/',
     'depends': ['base_phone', 'hr'],
-    'data': [
-        'security/ir.model.access.csv',
-        'hr_view.xml',
-        ],
-    'installable': False,
+    'data': ['security/ir.model.access.csv'],
+    'installable': True,
     'auto_install': True,
 }
