--- hr_phone/hr_view.xml
@@ -1,49 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<!--
-  © 2014-2016 Akretion (Alexis de Lattre <alexis.delattre@akretion.com>)
-  License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
--->
-
-<odoo>
-<data>
-
-
-<record id="view_employee_form" model="ir.ui.view">
-    <field name="name">hr_phone.hr_employee.form</field>
-    <field name="model">hr.employee</field>
-    <field name="inherit_id" ref="hr.view_employee_form"/>
-    <field name="arch" type="xml">
-        <field name="work_phone" position="attributes">
-            <attribute name="widget">phone</attribute>
-        </field>
-        <field name="mobile_phone" position="attributes">
-            <attribute name="widget">phone</attribute>
-        </field>
-    </field>
-</record>
-
-<record id="view_employee_tree" model="ir.ui.view">
-    <field name="name">hr_phone.hr_employee.tree</field>
-    <field name="model">hr.employee</field>
-    <field name="inherit_id" ref="hr.view_employee_tree"/>
-    <field name="arch" type="xml">
-        <field name="work_phone" position="attributes">
-            <attribute name="widget">phone</attribute>
-        </field>
-    </field>
-</record>
-
-<record id="view_partner_tree2" model="ir.ui.view">
-    <field name="name">hr_phone.hr.employee.tree2</field>
-    <field name="model">hr.employee</field>
-    <field name="inherit_id" ref="hr.view_partner_tree2"/>
-    <field name="arch" type="xml">
-        <field name="work_phone" position="attributes">
-            <attribute name="widget">phone</attribute>
-        </field>
-    </field>
-</record>
-
-
-</data>
-</odoo>
