+++ hr_phone/models/hr_employee.py
@@ -2,8 +2,8 @@
 # © 2012-2016 Akretion (Alexis de Lattre <alexis.delattre@akretion.com>)
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
 
-from openerp import models
-from openerp.addons.base_phone.fields import Phone
+from odoo import models
+from odoo.addons.base_phone.fields import Phone
 
 
 class HrEmployee(models.Model):
