+++ hr_recruitment_phone/__init__.py
@@ -1,3 +1,3 @@
 # -*- coding: utf-8 -*-
 
-from . import hr_recruitment_phone
+from . import models
