--- hr_recruitment_phone/hr_recruitment_view.xml
@@ -1,38 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<!--
-  © 2014-2016 Akretion (Alexis de Lattre <alexis.delattre@akretion.com>)
-  License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl)
--->
-
-<odoo>
-<data>
-
-
-<record id="crm_case_form_view_job" model="ir.ui.view">
-    <field name="name">hr_recruitment_phone.hr_applicant.form</field>
-    <field name="model">hr.applicant</field>
-    <field name="inherit_id" ref="hr_recruitment.crm_case_form_view_job"/>
-    <field name="arch" type="xml">
-        <field name="partner_phone" position="attributes">
-            <attribute name="widget">phone</attribute>
-        </field>
-        <field name="partner_mobile" position="attributes">
-            <attribute name="widget">phone</attribute>
-        </field>
-    </field>
-</record>
-
-<record id="crm_case_tree_view_job" model="ir.ui.view">
-    <field name="name">hr_recruitment_phone.hr_applicant.tree</field>
-    <field name="model">hr.applicant</field>
-    <field name="inherit_id" ref="hr_recruitment.crm_case_tree_view_job"/>
-    <field name="arch" type="xml">
-        <field name="partner_phone" position="attributes">
-            <attribute name="widget">phone</attribute>
-        </field>
-    </field>
-</record>
-
-
-</data>
-</odoo>
