+++ hr_recruitment_phone/models/__init__.py
@@ -1,0 +1,3 @@
+# -*- coding: utf-8 -*-
+
+from . import hr_applicant
