+++ hr_recruitment_phone/models/hr_applicant.py
@@ -2,8 +2,8 @@
 # © 2012-2016 Akretion (Alexis de Lattre <alexis.delattre@akretion.com>)
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
 
-from openerp import models, api
-from openerp.addons.base_phone.fields import Phone
+from odoo import models, api
+from odoo.addons.base_phone.fields import Phone
 
 
 class HrApplicant(models.Model):
